SE_BLOCK pattern moved from HW to IGNORED pattres
