+++ nncf/openvino/hardware/fused_patterns.py
@@ -147,48 +147,6 @@ def create_shift_scale() -> GraphPattern:
     return pattern
 
 
-@OPENVINO_HW_FUSED_PATTERNS.register(HWFusedPatternNames.SE_BLOCK)
-def create_se_block() -> GraphPattern:
-    pattern = GraphPattern()
-    any_node = pattern.add_node(
-        **{GraphPattern.LABEL_ATTR: "ANY", GraphPattern.METATYPE_ATTR: GraphPattern.NON_PATTERN_NODE_TYPE}
-    )
-    reduce_mean_node = pattern.add_node(
-        **{GraphPattern.LABEL_ATTR: "REDUCE_MEAN", GraphPattern.METATYPE_ATTR: om.OVReduceMeanMetatype}
-    )
-    linear_node_1 = pattern.add_node(
-        **{GraphPattern.METATYPE_ATTR: LINEAR_OPERATIONS, GraphPattern.LABEL_ATTR: "LINEAR"}
-    )
-    add_node_1 = pattern.add_node(**{GraphPattern.LABEL_ATTR: "ADD_BIAS", GraphPattern.METATYPE_ATTR: om.OVAddMetatype})
-    activation_node_1 = pattern.add_node(
-        **{
-            GraphPattern.LABEL_ATTR: "RELU, PRELU, SWISH",
-            GraphPattern.METATYPE_ATTR: [om.OVReluMetatype, om.OVPReluMetatype, om.OVSwishMetatype],
-        }
-    )
-    linear_node_2 = pattern.add_node(
-        **{GraphPattern.METATYPE_ATTR: LINEAR_OPERATIONS, GraphPattern.LABEL_ATTR: "LINEAR"}
-    )
-    add_node_2 = pattern.add_node(**{GraphPattern.LABEL_ATTR: "ADD_BIAS", GraphPattern.METATYPE_ATTR: om.OVAddMetatype})
-    activation_node_2 = pattern.add_node(
-        **{GraphPattern.LABEL_ATTR: "SIGMOID", GraphPattern.METATYPE_ATTR: om.OVSigmoidMetatype}
-    )
-    multiply_node = pattern.add_node(
-        **{GraphPattern.LABEL_ATTR: "MULTIPLY", GraphPattern.METATYPE_ATTR: om.OVMultiplyMetatype}
-    )
-
-    pattern.add_edge(any_node, reduce_mean_node)
-    pattern.add_edge(reduce_mean_node, linear_node_1)
-    pattern.add_edge(linear_node_1, add_node_1)
-    pattern.add_edge(add_node_1, activation_node_1)
-    pattern.add_edge(activation_node_1, linear_node_2)
-    pattern.add_edge(linear_node_2, add_node_2)
-    pattern.add_edge(add_node_2, activation_node_2)
-    pattern.add_edge(activation_node_2, multiply_node)
-    pattern.add_edge(any_node, multiply_node)
-    return pattern
-
-
 @OPENVINO_HW_FUSED_PATTERNS.register(HWFusedPatternNames.SOFTMAX_DIV)
 def create_softmax_div() -> GraphPattern:
     pattern = GraphPattern()
