+++ tests/common/test_pattern_manager.py
@@ -12,7 +12,6 @@
 from enum import Enum
 
 from nncf.common.graph.patterns.manager import PatternsManager
-from nncf.common.graph.patterns.patterns import AlgorithmType
 from nncf.common.graph.patterns.patterns import PatternDesc
 from nncf.common.utils.registry import Registry
 from nncf.parameters import ModelType
@@ -38,7 +37,7 @@ TEST_DEVICE_PATTERN_REGISTRY.register(DevicePatterns.COMMON_PATTERN)(None)
 def test_pattern_filter_device():
     manager = PatternsManager()
     filtered_patterns = manager._filter_patterns(
-        TEST_DEVICE_PATTERN_REGISTRY.registry_dict, device=TargetDevice.CPU, model_type=None, algorithm=None
+        TEST_DEVICE_PATTERN_REGISTRY.registry_dict, device=TargetDevice.CPU, model_type=None
     )
     assert len(filtered_patterns) == 2
     assert DevicePatterns.CPU_PATTERN in filtered_patterns
@@ -60,56 +59,14 @@ TEST_MODEL_TYPE_PATTERN_REGISTRY.register(ModelTypePatterns.COMMON_PATTERN)(None
 def test_pattern_filter_model_type():
     manager = PatternsManager()
     filtered_patterns = manager._filter_patterns(
-        TEST_MODEL_TYPE_PATTERN_REGISTRY.registry_dict, device=None, model_type=None, algorithm=None
+        TEST_MODEL_TYPE_PATTERN_REGISTRY.registry_dict, device=None, model_type=None
     )
     assert len(filtered_patterns) == 1
     assert ModelTypePatterns.COMMON_PATTERN in filtered_patterns
 
     filtered_patterns = manager._filter_patterns(
-        TEST_MODEL_TYPE_PATTERN_REGISTRY.registry_dict, device=None, model_type=ModelType.TRANSFORMER, algorithm=None
+        TEST_MODEL_TYPE_PATTERN_REGISTRY.registry_dict, device=None, model_type=ModelType.TRANSFORMER
     )
     assert len(filtered_patterns) == 2
     assert ModelTypePatterns.COMMON_PATTERN in filtered_patterns
     assert ModelTypePatterns.TRANSFORMER_PATTERN in filtered_patterns
-
-
-class AlgorithmTypePatterns(Enum):
-    QUANTIZATION_PATTERN = PatternDesc("QUANTIZATION_PATTERN", ignored_algorithms=[AlgorithmType.NAS])
-    NAS_PATTERN = PatternDesc("NAS_PATTERN", ignored_algorithms=[AlgorithmType.QUANTIZATION])
-    COMMON_PATTERN = PatternDesc("COMMON_PATTERN")
-    IGNORED_BY_BOTH = PatternDesc("IGNORED_BY_BOTH", ignored_algorithms=[AlgorithmType.QUANTIZATION, AlgorithmType.NAS])
-
-
-TEST_ALGORITHM_TYPE_PATTERN_REGISTRY = Registry("TEST_PATTERNS_REGISTRY")
-
-
-TEST_ALGORITHM_TYPE_PATTERN_REGISTRY.register(AlgorithmTypePatterns.QUANTIZATION_PATTERN)(None)
-TEST_ALGORITHM_TYPE_PATTERN_REGISTRY.register(AlgorithmTypePatterns.NAS_PATTERN)(None)
-TEST_ALGORITHM_TYPE_PATTERN_REGISTRY.register(AlgorithmTypePatterns.COMMON_PATTERN)(None)
-TEST_ALGORITHM_TYPE_PATTERN_REGISTRY.register(AlgorithmTypePatterns.IGNORED_BY_BOTH)(None)
-
-
-def test_pattern_filter_ignored_algorithm():
-    manager = PatternsManager()
-    filtered_patterns = manager._filter_patterns(
-        TEST_ALGORITHM_TYPE_PATTERN_REGISTRY.registry_dict, device=None, model_type=None, algorithm=None
-    )
-    assert len(filtered_patterns) == 4
-    assert all(pattern in filtered_patterns for pattern in AlgorithmTypePatterns)
-
-    filtered_patterns = manager._filter_patterns(
-        TEST_ALGORITHM_TYPE_PATTERN_REGISTRY.registry_dict,
-        device=None,
-        model_type=None,
-        algorithm=AlgorithmType.QUANTIZATION,
-    )
-    assert len(filtered_patterns) == 2
-    assert AlgorithmTypePatterns.QUANTIZATION_PATTERN in filtered_patterns
-    assert AlgorithmTypePatterns.COMMON_PATTERN in filtered_patterns
-
-    filtered_patterns = manager._filter_patterns(
-        TEST_ALGORITHM_TYPE_PATTERN_REGISTRY.registry_dict, device=None, model_type=None, algorithm=AlgorithmType.NAS
-    )
-    assert len(filtered_patterns) == 2
-    assert AlgorithmTypePatterns.NAS_PATTERN in filtered_patterns
-    assert AlgorithmTypePatterns.COMMON_PATTERN in filtered_patterns
